requirements.txt
howlongtobeatpy>=0.1.18
py-cord>=2.0.0b7
python-dotenv>=0.19.2